fix bug edit questions and answers, add and moves edit to myQuestions menu, add hooks and redirect
