+++ Client/src/router/index.js
@@ -2,6 +2,7 @@ import Vue from 'vue'
 import Router from 'vue-router'
 import Home from '@/components/home'
 import detailQuestion from '@/components/detailQuestion'
+import myQuestion from '@/components/myQuestion'
 
 Vue.use(Router)
 
@@ -13,10 +14,23 @@ export default new Router({
       component: Home
     },
     {
-      path: '/:id',
+      path: '/detailQuestion/:id',
       name: 'detailQuestion',
       component: detailQuestion,
       props: true
+    },
+    {
+      path: '/myQuestion',
+      name: 'myQuestion',
+      component: myQuestion,
+      beforeEnter: (to, from, next) => {
+        // console.log(from)
+        if (localStorage.getItem('token')) {
+          next()
+        } else {
+          next('/')
+        }
+      }
     }
   ]
 })
